Sleep for TIMEOUT if job was throttled
We sleep for TIMEOUT if no job was received (from redis, or queues list
was empty), thus it makes sense to sleep for same amount of time when
we got job but it was throttled.

This especially makes sense if you have a single queue that is hardly
throttled and has bazillions of jobs in it.