Add note about load order in sidekiq init file
When troubleshooting why throttling was not being applied, it was noticed that the call to `Sidekiq::Throttled.setup!` was happening at the top of `initializers/sidekiq.rb` in a rails app. This `.setup!` call was moved to the bottom of the sidekiq initializer file and throttle instantly starting working after restarting sidekiq.

Include a note about this configuration position in the README to prevent future headaches.