Replace "end" with "head" in test cases (#156)
`basic_fetch_spec.rb` tests the behavior of requeue of `retrieve_work`.
As far as I see, `requeue_throttled` uses Redis LPUSH, which inserts all
the specified values at the head of the list stored at key.

However, the previous test cases were telling
"pushes job back to the end queue". I think this is not correct.

When I was assessing this repository, I was a bit confused by the test
case, so I want to correct the test case.