doc: Add docs on using string key suffixes (#202)
Add docs about using string keys to create multiple limits per unit of
time.

I was confused about implementing this type of behavior until I saw your
comment in #149, explaining the use of strings for the key_suffix.
Hopefully this doc will help other realize they don't need a proc to do
this but that the key_suffix is required for having multiple limits.

I was simply doing this:

```ruby
throttle: [
  { limit: 500, period: 1.minute },
  { limit: 18_000, period: 1.day }
]
```

This showed up in the UI on Sidekiq Admin as two separate throttles so
it looked like it worked but it does not.

Signed-off-by: Jake Moffatt <jakeonrails@gmail.com>