Avoid race-condition when scheduling throttled jobs (#208)
When a job was throttled and the queue of pending jobs was fully drained
between the throttle? decision and the delay calculation, the retry code
would fail with "NoMethodError undefined method `last' for
nil:NilClass". I don't see a way to write a test for this since it is
very timing based and highly intermittent.

fixes #206

PS: I changed the appraisals file because install_if does all of the
bundling work regardless of whether the condition evaluates to true or
false. In fact, it fetches the packages list from gems.contribsys.com
before it even checks the condition. If BUNDLE_GEMS__CONTRIBSYS__COM
isn't set that fetch fails with an auth error.