backlog size aware throttling (#209)
This PR changes the `retry_in` calculation for concurrency throttled
jobs using the `:schedule` strategy. The approach is to estimate when a
throttled job will likely be runnable and schedule it for that time.
This estimate is based on the current backlog size for the job group
(estimated), the expected job runtime and the concurrency limit.

This should reduce/eliminate unnecessary delays in job processing caused
by the current "ignore throttled jobs for TTL seconds" approach while
retaining the benefits of the schedule strategy.


PS: I changed the appraisals file because install_if does all of the
bundling work regardless of whether the condition evaluates to true or
false. In fact, it fetches the packages list from gems.contribsys.com
before it even checks the condition. If BUNDLE_GEMS__CONTRIBSYS__COM
isn't set that fetch failed with an auth error.