Add spec showing bug with concurrency order
This spec fails if the dynamic rule is specified after the static
rule.

In this case when `throttle?` is called for the 4th job with the arg `99`
the static rule returns `false` and adds the job to the in progress set,
but the dynamic rule then returns `true`, so the job should not be
executed.

If the order the rules are configured is reversed then dynamic rule
will execute first, return `true`, and the static rule will never run
so the job will not be added to the in progress set.